+++ src/Data/TimeSeries/Stats.hs
@@ -1,6 +1,8 @@
 
 module Data.TimeSeries.Stats
     ( mean
+    , std
+    , variance
     ) where
 
 import Prelude hiding (max, min)
@@ -11,3 +13,14 @@ import Data.TimeSeries.Series (Series)
 mean :: Fractional a => Series a -> a
 mean xs = sum xs / fromIntegral (length xs)
 
+
+-- | Calculate Time Series variance
+variance :: Floating a => Series a -> a
+variance xs = sum (fmap (\x -> (x-mu)**2) xs) / fromIntegral (length xs)
+    where mu = mean xs
+
+
+
+-- | Calculate Time Series Standard Deviation
+std :: Floating a => Series a -> a
+std = sqrt . variance